# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/cars": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "List all cars",
                "description": "Retrieve a list of all cars with optional pagination parameters 'skip' and 'limit'.",
                "operationId": "0e2bcb5c69987da3dec0ff8fea117020",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Number of records to skip for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of records to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Car"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cars retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Car"
                ],
                "summary": "Create a new car",
                "description": "Create a new car record with the provided data.",
                "operationId": "2c89002b55d9b5bb9a87b4ae1c0cc39e",
                "requestBody": {
                    "description": "Car object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Car"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car saved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get a car by ID",
                "description": "Retrieve a single car by its ID.",
                "operationId": "97be050aefa7a40e66d5931d38ccea73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Car"
                ],
                "summary": "Update a car by ID",
                "description": "Update the details of an existing car by its ID.",
                "operationId": "66367a6f94de5301eeb824456a448f95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Car object with updated data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Car"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Car"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Car"
                ],
                "summary": "Delete a car by ID",
                "description": "Delete an existing car by its ID.",
                "operationId": "ac6f0c58d20a6ccd0b1f4ae539f433bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Car deleted successfully"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "title": "Car",
                "description": "A Car model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "car_name": {
                        "type": "string",
                        "example": "Avanza"
                    },
                    "day_rate": {
                        "type": "number",
                        "format": "double",
                        "example": 150
                    },
                    "month_rate": {
                        "type": "number",
                        "format": "double",
                        "example": 3000
                    },
                    "image": {
                        "description": "URL or filename of the image",
                        "type": "string",
                        "example": "avanza.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}